Put CalendarDay into a separate component
