+++ src/components/TodosCalendar/TodosCalendar.jsx
@@ -1,34 +1,11 @@
-import { Badge } from '@mui/material';
-import { DateCalendar, LocalizationProvider, PickersDay } from '@mui/x-date-pickers';
+import { DateCalendar, LocalizationProvider } from '@mui/x-date-pickers';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { useDispatch } from "react-redux";
 import { setDate } from "../../features/filtersSlice";
+import { CalendarDay } from '../../components';
 
 
 export function TodosCalendar() {
-    const highlightedDay = (pickersDayProps) => {
-        // daysWithTodos shouldn't be passed to PickerDays (React doesn't recognise it as a correct prop)
-        const {daysWithTodos, day, outsideCurrentMonth, ...otherProps} = pickersDayProps;
-
-        // Now I just need to compare the date of this day to a list of days with a todo
-        const hasTodo = daysWithTodos.some(dayWithTodo => dayWithTodo === `${day.year()}-${day.month() + 1}-${day.date()}`);
-        const isHighlighted = !outsideCurrentMonth && hasTodo;
-
-        return (
-            <Badge 
-                variant={isHighlighted && "dot"}
-                overlap="circular"
-                color='primary'
-            >
-                <PickersDay
-                    day={day}
-                    outsideCurrentMonth={outsideCurrentMonth}
-                    {...otherProps}
-                />
-            </Badge>
-        );
-    }
-
     const dispatch = useDispatch();
 
     const handleDateSelection = (date) => {
@@ -41,7 +18,7 @@ export function TodosCalendar() {
         <LocalizationProvider dateAdapter={AdapterDayjs} adapterLocale="en-gb">
             <DateCalendar
                 onChange={handleDateSelection}
-                slots={{day: highlightedDay}}
+                slots={{day: CalendarDay}}
                 slotProps={{day: {daysWithTodos}}}
             >
             </DateCalendar>
