+++ src/components/index.js
@@ -4,4 +4,5 @@ export { TodoDetails } from './TodoDetails/TodoDetails';
 export { TodoItem } from './TodoItem/TodoItem';
 export { CalendarBlock } from './CalendarBlock/CalendarBlock';
 export { TodoBlock } from './TodoBlock/TodoBlock';
-export { TodosCalendar } from './TodosCalendar/TodosCalendar'
+export { TodosCalendar } from './TodosCalendar/TodosCalendar';
+export { CalendarDay } from './CalendarDay/CalendarDay';
